+++ app/Http/Controllers/QuestionsController.php
@@ -11,8 +11,10 @@ use App\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Auth;
+use Validator;
 use Illuminate\Auth\AuthManager;
 use Illuminate\Support\Facades\Crypt;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 
@@ -69,6 +71,9 @@ class QuestionsController extends Controller
 
     public function index()
     {
+        if (Gate::denies('admin')) {
+            return redirect()->route('dashboard');
+        }
         $questions = $this->questions->all();
         return view('questions/index', ['questions' => $questions]);
     }
@@ -76,6 +81,9 @@ class QuestionsController extends Controller
 
     public function create()
     {
+        if (Gate::denies('admin')) {
+            return redirect()->route('dashboard');
+        }
         $datas = $this->activities->all();
 //        return $datas;
         return view('questions/create', ['datas' => $datas]);
@@ -85,17 +93,58 @@ class QuestionsController extends Controller
     public function store(Request $request)
     {
 
+//        Validator::make($request->all(), [
+//            'activities_id' => 'required',
+//            'content_of_question' => 'required',
+//            'a' => 'required',
+//            'b' => 'required',
+//            'c' => 'required',
+//            'd' => 'required',
+//            'image' => 'mimes:jpeg,png',
+//            'image_r' => 'mimes:jpeg,png',
+//            //'tip' => 'required',
+////            'answer' => 'required',
+//            'answer' => [
+//                'required',
+//                function ($attribute, $value, $fail) {
+//                    if (preg_match("/[a-dA-D]/", $value)!==1) {
+//                        $fail('A resposta deve conter letras de A - D');
+//                    }
+//                },
+//
+//            ],
+//        ])->validate();
+        // Todo resolver problemas de validação
+        $rules = [
+            'activities_id' => 'required',
+            'content_of_question' => 'required',
+            'a' => 'required',
+            'b' => 'required',
+            'c' => 'required',
+            'd' => 'required',
+            'image' => 'mimes:jpeg,png',
+            'image_r' => 'mimes:jpeg,png',
+            //'tip' => 'required',
+            'answer' => 'required',
+        ];
+        $validatedData = $request->validate($rules);
+
+        if (Gate::denies('admin')) {
+            return redirect()->route('dashboard');
+        }
+
         $data = $request->all();
         $date = $this->carbon->now()->format('d-m-Y');
 
         $path = Storage::put($date, $request->file('image'), 'public');
+        $path2 = Storage::put($date, $request->file('image_r'), 'public');
 
 
         if (Storage::disk('s3')->exists($path)) {
 
             $url = 'https://s3-sa-east-1.amazonaws.com/guitterquiz/' . $path;
             // TODO fazer importação da imagem de resposta
-            $url_r = 'https://s3-sa-east-1.amazonaws.com/guitterquiz/' . $path;
+            $url_r = 'https://s3-sa-east-1.amazonaws.com/guitterquiz/' . $path2;
             $url_r = Crypt::encryptString($url_r);
             $answer = Crypt::encryptString(strtolower($request->input('answer')));
 
@@ -123,24 +172,36 @@ class QuestionsController extends Controller
 
     public function show($id)
     {
+        if (Gate::denies('admin')) {
+            return redirect()->route('dashboard');
+        }
         //
     }
 
 
     public function edit($id)
     {
+        if (Gate::denies('admin')) {
+            return redirect()->route('dashboard');
+        }
         return view('questions/edit');
     }
 
 
     public function update(Request $request, $id)
     {
+        if (Gate::denies('admin')) {
+            return redirect()->route('dashboard');
+        }
         //
     }
 
 
     public function destroy($id)
     {
+        if (Gate::denies('admin')) {
+            return redirect()->route('dashboard');
+        }
          $result = $this->questions->delete($id);
         if ($result) {
             return redirect()->route('questions');
@@ -155,8 +216,14 @@ class QuestionsController extends Controller
 
     public function showquestions()
     {
+
         $date = $this->carbon->now()->format('Y-m-d');
         $data = $this->activities->findWhere(['date' => $date])->toArray();
+
+        if(sizeof($data) == 0){
+            return redirect()->route('dashboard');
+            // TODO colocar mensagem de nenhum questionario disponivel
+        }
 
         $performance = $this->authManager->user()->performance;
 
@@ -192,6 +259,7 @@ class QuestionsController extends Controller
     {
         $date = $this->carbon->now()->format('Y-m-d');
         $length = $request->input('length');
+        $discipline_name = $request->input('discipline_name');
         $total = 0;
         if ($length) {
             $result = [];
@@ -253,12 +321,13 @@ class QuestionsController extends Controller
         }
 
         $note = ($total / $length) * 100;
-        $to_save = ['note' => $note, 'date' => $date, 'resolved' => '1'];
+        $to_save = ['note' => '', 'date' => $date, 'hit_percentage' => $note, 'number_of_questions'=>$length,'discipline_name'=>$discipline_name,'total_resolved'=>'','resolved'=>$total];
 
         $this->authManager->user()->performance()->create($to_save);
-
         return view('questions/result', ['result' => $result, 'length' => $length, 'total' => $total]);
 
 
     }
+
+
 }
